Scientific Notation added

+++ src/ast/expression/float_literal.rs
@@ -3,6 +3,6 @@ use crate::token::Token;
 #[derive(Default, PartialOrd, Ord, Clone, PartialEq, Eq, Hash, Debug)]
 pub struct FloatLiteral {
     pub token: Token,
-    pub integer_part: i32,
+    pub integer_part: i64,
     pub float_part: u64,
 }
